requirements.txt
# Core dependencies
setuptools>=69.0.0,<75.0.0
python-dotenv>=1.0.0,<2.0.0
requests>=2.32.0,<3.0.0
pandas>=2.2.0,<3.0.0
numpy>=1.26.0,<3.0.0
scikit-learn>=1.5.0,<2.0.0
scipy>=1.10.0,<2.0.0
pydantic>=2.8.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
fastapi>=0.115.0,<1.0.0
typing-extensions>=4.8.0,<5.0.0
uvicorn>=0.30.0,<1.0.0
sqlalchemy>=2.0.23,<2.1.0
dependency-injector>=4.40.0,<5.0.0
email-validator>=2.0.0,<3.0.0
asyncpg>=0.29.0,<1.0.0
redis>=5.0.1,<6.0.0
websockets>=10.3,<14.0
aiohttp>=3.9.0,<4.0.0
pyjwt>=2.9.0,<3.0.0
bcrypt>=4.1.2,<5.0.0
passlib>=1.7.4,<2.0.0
python-multipart>=0.0.6,<1.0.0

# ML/AI — on x86 these pull CPU-only builds; on ARM64 they're CPU-only by default
tensorflow>=2.15.0,<2.18.0
torch>=2.0.0,<3.0.0
transformers>=4.35.0,<5.0.0
xgboost>=2.0.0,<3.0.0

# NLP and Sentiment Analysis
textblob>=0.17.1,<1.0.0
vaderSentiment>=3.3.2,<4.0.0
nltk>=3.8.1,<4.0.0

# Financial Data
yfinance>=0.2.18,<1.0.0
alpha-vantage>=2.3.0,<4.0.0
polygon-api-client>=1.13.0,<2.0.0
finnhub-python>=2.4.14,<3.0.0

# Database Migrations
alembic>=1.13.1,<2.0.0
psycopg2-binary>=2.9.9,<3.0.0

# Technical Analysis
ta>=0.11.0,<1.0.0

# AI Explainability
shap>=0.44.0,<1.0.0

# AI Chat
anthropic>=0.40.0,<1.0.0

# AWS
boto3>=1.35.0,<2.0.0

# Utilities
python-slugify>=8.0.1,<9.0.0
cryptography>=42.0.0,<44.0.0

# Rate Limiting
slowapi>=0.1.8,<1.0.0

# Scheduling
apscheduler>=3.10.0,<4.0.0

# Observability (2026 rules §6 — tracing, metrics, structured logs)
opentelemetry-api>=1.27.0,<2.0.0
opentelemetry-sdk>=1.27.0,<2.0.0
opentelemetry-instrumentation-fastapi>=0.48b0,<1.0.0
opentelemetry-instrumentation-sqlalchemy>=0.48b0,<1.0.0
opentelemetry-instrumentation-httpx>=0.48b0,<1.0.0
opentelemetry-exporter-otlp-proto-http>=1.27.0,<2.0.0
prometheus-fastapi-instrumentator>=7.0.0,<8.0.0

# MCP — bounded-context servers (2026 rules §3.5)
# jsonschema validates tool input/output at every boundary.
jsonschema>=4.17.0,<5.0.0

# Live trading — per-user TOTP for 2FA on live-mode enable + per-order challenge
# (ADR-002). Pinned to the current 2.x stable.
pyotp>=2.9.0,<3.0.0
